/api/adminの認証追加

+++ src/app/admin/categories/[id]/page.tsx
@@ -5,21 +5,28 @@ import { useRouter } from 'next/navigation'
 import type { Category, PostDetailsProps } from '@/app/_types/index'
 import CategoryForm from '@/app/admin/categories/_components/CategoryForm'
 import { useCategory } from '@/app/_hooks/useCategory'
+import { useSupabaseSession } from '@/app/_hooks/useSupabaseSession'
 
 const EditCategory: React.FC<PostDetailsProps> = ({ params }) => {
   const { id } = params
   const { category: fetchedCat, isLoading, isError } = useCategory(id)
   const router = useRouter()
   const [category, setCategory] = useState<Category | null>(null)
+  const { token } = useSupabaseSession()
 
   useEffect(() => {
     setCategory(fetchedCat ?? null)
   },[fetchedCat])
 
   const handleEdit = async () => {
+    if (!token) return 
+
     const response: Response = await fetch(`${process.env.NEXT_PUBLIC_APP_BASE_URL}/api/admin/categories/${id}`, {
       method: 'PUT',
-      headers: {'Content-Type': 'application/json'},
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': token,
+      },
       body: JSON.stringify({
         name: category?.name,
         posts: category?.posts.map(elem => ({id: elem.postId}))
@@ -35,8 +42,14 @@ const EditCategory: React.FC<PostDetailsProps> = ({ params }) => {
   }
 
   const handleDelete = async () => {
+    if (!token) return 
+    
     const response: Response = await fetch(`${process.env.NEXT_PUBLIC_APP_BASE_URL}/api/admin/categories/${id}`, {
       method: 'DELETE',
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': token,
+      },
       })
 
     if (response.ok) {
